requirements.txt
Django==1.6.6
South==1.0
dj-database-url==0.3.0
django-crispy-forms==1.4.0
django-storages==1.1.8
honcho>=0.5.0
gunicorn==19.1.1
jedi==0.8.1-final0
wsgiref==0.1.2
django-widget-tweaks==1.3
django-eztables==0.3.2
